Cover the spawn failure paths the file split left ungated
Moving read_file/write_file into file_runtime.c took ~50 well-covered lines
out of system_runtime.c's denominator, so the remainder fell from 70.24% to
67.35% on Linux and the C ratchet failed at its 68% floor. The threshold is
correct; the coverage was missing. Three paths that nothing exercised now
have tests:

- A child killed by a SIGNAL must report -1. WEXITSTATUS of a signalled
  status is the signal number, so an unchecked read makes SIGKILL
  indistinguishable from `exit 9`. The command signals the tracked process
  itself (`kill -9 $$`): the runtime execs `/bin/sh -c`, so a nested `sh -c`
  would only work where the outer shell exec's the inner one -- bash does,
  dash does not, and on Linux the tracked shell exits NORMALLY with 137.
- pipe() and popen() denied a descriptor, via an RLIMIT_NOFILE capped at the
  lowest free fd. No argument can provoke these branches, and they are the
  ones that run when a long-lived program exhausts its table.
- The MAX_PROCESSES ceiling. next_process_id only increments -- cleanup
  frees the slot but never returns the id -- so a program that has spawned
  MAX_PROCESSES times can never spawn again with every slot free. Reached
  without a fork, since the descriptor cap fails each attempt at the pipe.

Measured on both platforms the gate runs on: macOS 74.86% -> 79.89%, Linux
73.91% -> 79.50% (gcc 13, same 9 lines). Threshold left at 68 until CI
reports its own Linux number, then ratcheted against a measurement rather
than a prediction.

The file tests move to test_file_runtime.c with the source they cover:
test_system_runtime.c was 503 lines, over the size budget, and one suite per
translation unit is what the split already implied.
